Move Ionization methods from FusionRuntime to TopLevel.
Fixes #515
