+++ runtime/src/main/java/dev/ionfusion/fusion/StandardTopLevel.java
@@ -10,6 +10,9 @@ import static dev.ionfusion.fusion.StandardReader.readSyntax;
 import static dev.ionfusion.runtime._private.util.Ordinals.friendlyIndex;
 import static dev.ionfusion.runtime.base.ModuleIdentity.isValidAbsoluteModulePath;
 
+import com.amazon.ion.IonValue;
+import com.amazon.ion.ValueFactory;
+
 import com.amazon.ion.IonReader;
 import com.amazon.ion.IonWriter;
 import dev.ionfusion.runtime._private.cover.CoverageCollector;
@@ -339,4 +342,23 @@ final class StandardTopLevel
             return null;
         });
     }
+
+
+    @Override
+    public IonValue ionize(Object fusionValue, ValueFactory factory)
+        throws FusionException
+    {
+        return withEvaluator(eval ->
+                                 FusionValue.copyToIonValue(eval, fusionValue, factory));
+    }
+
+
+    @Override
+    public IonValue ionizeMaybe(Object fusionValue, ValueFactory factory)
+        throws FusionException
+    {
+        return withEvaluator(eval ->
+                                 FusionValue.copyToIonValueMaybe(eval, fusionValue, factory));
+    }
+
 }
